Factor out the computed substances table
